complete reset of database models + migrations
